+++ src/components/atoms/modals/AuthButton/style.ts
