+++ src/components/atoms/modals/ModalContentWrapper/ModalContentWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode } from 'react';
 import * as S from './style';
-import ModalHeader from '../../modules/ModalHeader/ModalHeader';
-import ModalButtonWrapper from '../../modules/ModalButtonWrapper/ModalButtonWrapper';
+import ModalHeader from '../../../modules/ModalHeader/ModalHeader';
+import ModalButtonWrapper from '../../../modules/ModalButtonWrapper/ModalButtonWrapper';
 
 interface Props {
   isLogin?: boolean;
